==> Notebook 2 - Exploratory Analysis.ipynb <==
Import the necessary packages.

import os
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt 

Set the path to where the munged data is located.

path_to_data = '/home/bsknight/Documents/Personal_Training_Git/Udacity/Udacity-Machine_Learning_Nanodegree/' + \
               'Capstone_Project/Data' 
os.chdir(path_to_data)

Read in the data.

df = pd.read_csv('./munged_df.csv', index_col='account_id')

df.head(5)

What is the size of the data set?

print 'Each observation is an indivual account'
print 'Our data set has ' + str(len(df)) + ' observations.'
print 'Our data set has ' + str(len(list(df)) - 1) + ' features.'
print 'The probability of a label designating a sucessful conversion to paying customer is ' + \
     str(df['cc'].sum() / len(df)) + '.'

The labels for our data are contained in the 'cc' column. Let's see how frequent the labels are.

Create a horizontal bar chart.

plt.figure(figsize=(8,3))
a = pd.DataFrame(pd.Series(df["cc"]).value_counts()).iloc[0]['cc']
b = pd.DataFrame(pd.Series(df["cc"]).value_counts()).iloc[1]['cc']
plt.barh([0], [a], label="No Credit Card Added", color='grey', align='center', alpha=0.5)
plt.barh([1], [b], label="Credit Card Added", color = 'green', align='center', alpha=0.5)
y_labels = ["No Credit Card Added", "Credit Card Added"]
x_values = np.arange(0,16001,1000)
x_labels= ['0', '1,000', '2,000', '3,000', '4,000', '5,000', '6,000', '7,000', '8,000', '9,000', '10,000', 
           '11,000', '12,000', '13,000', '14,000', '15,000', '16,000']
plt.xticks(x_values, x_labels, rotation=45)
plt.yticks([0,1], y_labels)
plt.ylabel('Account Status', rotation='horizontal', fontsize=13)
plt.xlabel('# of Accounts Created (November 2015 to January 2017)', fontsize=13)
# plt.title('The Observation Labels: Conversion to Paying Account', fontsize=18, y=1.01, x=0.3)
plt.tight_layout()
# plt.legend()
# plt.show()
path = '/home/bsknight/Documents/Personal_Training_Git/Udacity/Udacity-Machine_Learning_Nanodegree/' + \
       'Capstone_Project/Images'
savepath = os.path.join(path, 'exploratory_analysis-labels.png')
plt.savefig(savepath)
plt.clf()

Create a pie chart.

# plt.figure(figsize=(16,7.5))
# plt.rcParams['font.size'] = 24.0
# labels = 'Converted to \n Paying Customer \n (Added a Credit Card)', 'Never Converted  \n (No Credit Card Added)'
# sizes = [df["cc"].sum(), (len(df)-df["cc"].sum()) ]
# colors = ['g', '#7F7F7F']
# explode = (0.0, 0.4)  # explode 1st slice

# plt.pie(sizes, explode=explode, labels=labels, colors=colors,
#         autopct='%1.1f%%', shadow=True, startangle=140)
# # plt.title('City:[3,500, 6,800 Students] vs. Metro:[6,800, 712,131 Students]', 
# #           y=1, fontsize=20)
# # plt.suptitle('District Segementation', y=0.98, fontsize=25)
# # plt.show()

# plt.axis('equal')
# # plt.show()
# path = '/home/bsknight/Documents/Personal_Training_Git/Udacity/Udacity-Machine_Learning_Nanodegree/' + \
#        'Capstone_Project/Images'
# savepath = os.path.join(path, 'exploratory_analysis-labels.png')
# plt.savefig(savepath)
# plt.clf()

Now that we've examined the distrubution of labels, lets calculate the summary statistics for the other features. 

means = pd.DataFrame(df.mean())
means = means.reset_index()
means = means.rename(columns={'index': 'Feature', 0: 'Mean'})
means.sort_values(by='Mean', ascending=0, inplace=True)
bins = np.arange(0, 30, 1)
plt.hist(means["Mean"], bins, histtype='bar', alpha=0.5)
plt.xlabel('Feature Mean')
plt.ylabel('Number of Features')
plt.title('Histogram of Feature Means (Bin = 1)')
plt.tight_layout()
# plt.ylim((0,30))
# plt.show()
savepath = os.path.join(path, 'exploratory_analysis-feature_means.png')
plt.savefig(savepath)
plt.clf()

Now for the standard deviations.

sds = pd.DataFrame(df.std())  
sds = sds.reset_index()
sds = sds.rename(columns={'index': 'Feature', 0: 'SD'})
sds.sort_values(by='SD', ascending=0, inplace=True)
bins = np.arange(0, 200, 10)
plt.hist(sds["SD"], bins, histtype='bar', alpha=0.5)
plt.xlabel('Feature Standard Deviation')
plt.ylabel('Number of Features')
plt.title('Histogram of Feature Standard Deviations (Bin = 10)')
plt.tight_layout()
# plt.ylim((0,100))
# plt.show()
savepath = os.path.join(path, 'exploratory_analysis-feature_sds.png')
plt.savefig(savepath)
plt.clf()
